test cucumber + parquet : renaming modeler_output_file.py into invest_problems.py
